bean: Data Thinning of Species Occurrences in Environmental Space

Provides a suite of tools to mitigate sampling bias in species occurrence records by thinning data in the environmental space (E-space). This process could help increase accuracy and precision in species distribution modeling (SDM, or ecological niche modeling, ENM). The package offers a data-driven protocol to determine thinning parameters. Thinning methods (stochastic and deterministic) help reduce oversampled conditions and outlier observations. The name 'bean' reflects the core principle of the method: ensuring that each “pod” (a grid cell in E-space) contains only a specified number of “beans” (occurrence points).

See also

Author

Maintainer: Paanwaris Paansri paanwaris@vt.edu

Authors:

Paanwaris Paansri paanwaris@vt.edu

Other contributors:

Luis E. Escobar escobar1@vt.edu [contributor]